refactor(adapter): extract shared spawn helper for CLI agents
Move SpawnedProcess/SpawnFn/defaultSpawn out of opencode-adapter.ts into a
reusable adapters/spawn.ts so the OpenCode and upcoming Claude Code adapters
share one shell-free, stdin-closed spawn surface. opencode-adapter.ts re-exports
the types for backwards-compatible imports.
